fix(EditTask):Fixed Sprint Data Changes
